pyproject.toml
[project]
name = "acstreamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pywebview>=6.1",
    "streamlit>=1.54.0",
]
